pyproject.toml
[tool.poetry]
name = "flask-contacts"
version = "0.1.0"
description = ""
authors = ["Leo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_contacts"}]

[tool.poetry.dependencies]
python = "^3.10"
flask-sqlalchemy = "^3.0.2"
flask-mail = "^0.9.1"
bootstrap-flask = "^2.2.0"
flask = "^2.2.2"
flask-restful = "^0.3.9"
python-decouple = "^3.6"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"
flask-migrate = "^4.0.0"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
ward = "^0.67.0b0"
splinter = {version = "^0.18.1", extras = ["flask"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
